cleanup and reorg
